fix(wasm): close the channels and value shapes adversarial review found (ARN-243)
Grok's first adversarial pass found four real gaps in the gate, all fixed here.

A fifth guest telemetry channel: host_emit_metric takes guest-named, guest-valued
string tags that reach both the OTel meter and a span event, ungated. It now takes
the same rule, redacted before either sink reads the tags.

The clamp and the allowlist disagreed about the namespace. The allowlist
normalized the key, the clamp matched the raw one, so GEN_AI.request.model passed
as recognised metadata and then skipped the clamp — carrying a whole prompt. Every
namespace test now goes through is_llm_namespace_key.

The clamp only matched strings, so gen_ai.request.model = ["<the whole prompt>"]
rode through untouched. Structured values under a metadata key are now dropped —
no key in the allowlist is structured — while numbers and booleans still pass.

Dispatch was still a denylist: a sink added later that read _gen_ai_completion
would have found it in the map. It is now an allowlist over the _gen_ai_ /
_dd_llmobs_ prefixes, consistent with the other channels.

Two testing gaps it also found: deleting the wide-event call site left every test
green (the unit tests call the helper, not the call site), and the guest-span
filter ran in two places so a mutation to either was masked by the other. There is
now one place guest attributes are judged, and a source contract on the
wide-event and metric call sites. host_start_span names are clamped like hint
names.

Each fix mutation-checked: structured values passed through, span name unclamped,
dispatch reverted to a denylist, wide-event call deleted, metric redaction deleted
or moved after its sink, raw-key clamp restored — all turn the suite red.

Test modules extracted from dispatch/wasm.rs, span_hints.rs, authorized_host.rs
and a redaction submodule split out of guest_spans.rs, so the readability ratchet
passes without moving its baseline.
